document and refactor
* lots of @overwrite annotations
* add some documentation
* remove performancewatcher -- should be better integrated into checkpointing
* use diamond operator for new collections

+++ src/test/java/at/ac/ait/archistar/integration/AbstractIntegrationTest.java
@@ -55,7 +55,7 @@ public abstract class AbstractIntegrationTest {
              */
             int increaseCount = 0;
             for (Map.Entry<String, Integer> m : serverConfig.getStorageFragmentCounts().entrySet()) {
-                int oldValue = fragCount.get(m.getKey()).intValue();
+                int oldValue = fragCount.get(m.getKey());
 
                 if (m.getValue() == (oldValue + 1)) {
                     increaseCount++;
